Refactor FriendsList component
